Move `config` to `internal/config`
